Package all pyxis dependencies

+++ specs/pyxis/cluster-init/scripts/20-pyxis.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 set -e
+set -x
+
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source "$script_dir/../files/common.sh" 
 
@@ -21,35 +23,48 @@ function install_pyxis_library() {
 
 function build_pyxis() {
 
-   # Check and install 'make' based on the OS
-   if [[ -f /etc/os-release ]]; then
-       . /etc/os-release
-       if [[ "$ID" == "ubuntu" ]]; then
-           logger -s "Installing 'make' on Ubuntu"
-           apt update && apt install -y make gcc wget
-       else
-           logger -s "Installing 'make' on non-Ubuntu system (assuming RHEL/CentOS)"
-           yum install -y make gcc wget
-       fi
-   else
-       logger -s "Unable to detect OS. Please ensure 'make' is installed."
-       exit 1
-   fi
+    BLOB_FILE=pyxis-artifacts.tar.gz
+    if [[ ! -f $PYXIS_DIR/spank_pyxis.so ]]; then
 
-   logger -s "Downloading Pyxis source code $PYXIS_VERSION"
+        # Check and install 'make' based on the OS
+        if [[ -f /etc/os-release ]]; then
+            . /etc/os-release
+            if [[ "$ID" == "ubuntu" ]]; then
+                logger -s "Installing 'make' on Ubuntu"
+                apt update && apt install -y make gcc wget
+            else
+                logger -s "Installing 'make' on non-Ubuntu system (assuming RHEL/CentOS)"
+                yum install -y make gcc wget
+            fi
+        else
+            logger -s "Unable to detect OS. Please ensure 'make' is installed."
+            exit 1
+        fi
 
-   cd /tmp
-   wget -q https://github.com/NVIDIA/pyxis/archive/refs/tags/v$PYXIS_VERSION.tar.gz
-   tar -xzf v$PYXIS_VERSION.tar.gz
-   cd pyxis-$PYXIS_VERSION
-   logger -s "Building Pyxis"
-   make
+        logger -s "Downloading Pyxis source code $PYXIS_VERSION"
 
-   # Copy pyxis library to /opt/pyxis
-   logger -s "Copying Pyxis library to $PYXIS_DIR"
-   mkdir -p $PYXIS_DIR
-   cp -fv spank_pyxis.so $PYXIS_DIR
-   chmod +x $PYXIS_DIR/spank_pyxis.so
+        if [[ -d /tmp/pyxis-artifacts ]]; then
+            cd /tmp/pyxis-artifacts
+        else
+            cd /tmp
+            project_name=$(jetpack config slurm.project_name slurm)
+            jetpack download --project $project_name $BLOB_FILE
+            tar -xvf $BLOB_FILE
+            cd /tmp/pyxis-artifacts
+        fi
+        tar -xzf pyxis-$PYXIS_VERSION.tar.gz
+        cd pyxis-$PYXIS_VERSION
+        logger -s "Building Pyxis"
+        make
+
+        # Copy pyxis library to /opt/pyxis
+        logger -s "Copying Pyxis library to $PYXIS_DIR"
+        mkdir -p $PYXIS_DIR
+        cp -fv spank_pyxis.so $PYXIS_DIR
+        chmod +x $PYXIS_DIR/spank_pyxis.so
+    else
+        echo "Pyxis already installed"
+    fi
 }
 
 logger -s "Install Pyxis library"
